"java tests and moving ui files"
