+++ backend/src/test/java/com/cloudera/cai/rag/external/RagBackendClientTest.java
@@ -48,6 +48,7 @@ import com.cloudera.cai.util.SimpleHttpClient.TrackedHttpRequest;
 import com.cloudera.cai.util.Tracker;
 import com.cloudera.cai.util.exceptions.ClientError;
 import com.cloudera.cai.util.exceptions.NotFound;
+import com.cloudera.cai.util.exceptions.ServerError;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpMethod;
@@ -107,6 +108,35 @@ class RagBackendClientTest {
   }
 
   @Test
+  void indexFile_serverError() {
+    Tracker<TrackedHttpRequest<?>> tracker = new Tracker<>();
+    RagBackendClient client =
+        new RagBackendClient(
+            SimpleHttpClient.createNull(
+                tracker, new ServerError("{\"detail\": \"things are bad\"}", 500)));
+    IndexConfiguration indexConfiguration = new IndexConfiguration(123, 2);
+    RagDocument document = indexRequest("documentId", "s3Path", 1234L, "myfile.mp3");
+
+    assertThatThrownBy(() -> client.indexFile(document, "bucketName", indexConfiguration))
+        .isInstanceOf(ServerError.class)
+        .hasMessage("things are bad");
+
+    List<TrackedHttpRequest<?>> values = tracker.getValues();
+    assertThat(values)
+        .hasSize(1)
+        .contains(
+            new TrackedHttpRequest<>(
+                HttpMethod.POST,
+                "http://rag-backend:8000/data_sources/"
+                    + 1234L
+                    + "/documents/"
+                    + "documentId"
+                    + "/index",
+                new RagBackendClient.IndexRequest(
+                    "bucketName", "s3Path", "myfile.mp3", indexConfiguration)));
+  }
+
+  @Test
   void createSummary() {
     Tracker<TrackedHttpRequest<?>> tracker = new Tracker<>();
     RagBackendClient client = new RagBackendClient(SimpleHttpClient.createNull(tracker));
@@ -136,6 +166,29 @@ class RagBackendClientTest {
     assertThatThrownBy(() -> client.createSummary(document, "bucketName"))
         .isInstanceOf(ClientError.class)
         .hasMessage("Unsupported media type");
+
+    List<TrackedHttpRequest<?>> values = tracker.getValues();
+    assertThat(values)
+        .hasSize(1)
+        .contains(
+            new TrackedHttpRequest<>(
+                HttpMethod.POST,
+                "http://rag-backend:8000/data_sources/1234/documents/" + "documentId" + "/summary",
+                new RagBackendClient.SummaryRequest("bucketName", "s3Path", "myfile.pdf")));
+  }
+
+  @Test
+  void createSummary_serverError() {
+    Tracker<TrackedHttpRequest<?>> tracker = new Tracker<>();
+    RagBackendClient client =
+        new RagBackendClient(
+            SimpleHttpClient.createNull(
+                tracker, new ServerError("{\"detail\": \"things are bad\"}", 500)));
+    RagDocument document = indexRequest("documentId", "s3Path", 1234L, "myfile.pdf");
+
+    assertThatThrownBy(() -> client.createSummary(document, "bucketName"))
+        .isInstanceOf(ServerError.class)
+        .hasMessage("things are bad");
 
     List<TrackedHttpRequest<?>> values = tracker.getValues();
     assertThat(values)
